models forlder created and model classes are moved there
