+++ LessonMonitor/LessonMonitor.Api/Models/RoadMapModel.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace LessonMonitor.Api.Models
+{
+    public class RoadMapModel
+    {
+    }
+}
