+++ LessonMonitor/LessonMonitor.Api/Models/Skill.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace LessonMonitor.Api
+namespace LessonMonitor.Api.Models
 {
     public class Skill
     {
